pyproject.toml
[tool.poetry]
name = "training_canvas"
version = "0.1.0"
description = "A training tool for a simple classification task, where users draw objects, and a CNN is trained in the background to classify this object"
authors = ["Julia Sammer <juliasammer@gmx.at>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
SQLAlchemy = "^2.0.4"
alembic = "^1.9.4"
fastapi = "^0.92.0"
pydantic = "^1.10.5"
uvicorn = "^0.20.0"
psycopg2-binary = "^2.9.5"
azure-storage-blob = "^12.15.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.0.0"
pylint = "^2.16.1"
pytest = "^7.2.1"
black = "^23.1.0"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
